+++ src/main/java/com/gobongbob/festamate/global/config/webSocket/WebSocketInterceptor.java
@@ -1,4 +1,4 @@
-package com.gobongbob.festamate.global.config;
+package com.gobongbob.festamate.global.config.webSocket;
 
 import com.gobongbob.festamate.domain.auth.jwt.domain.CustomMemberDetails;
 import com.gobongbob.festamate.domain.member.persistence.MemberRepository;
